fix: address three Codex review findings on PR #1221
- service_compare_pipeline.py: move ResolvedComparePair.resolved_env_matrix
  to the end of the dataclass field list, after resolved_execution_context.
  The prior insertion point silently repointed a positional 7+-arg Tier-2
  caller's execution-context value onto resolved_env_matrix instead of
  resolved_execution_context. All real construction sites already use
  keyword arguments, so this reorder changes no runtime behavior; it only
  removes the latent positional-argument hazard for a future caller.

- docs/reference/python-api-reference.md: regenerated via
  `python scripts/gen_python_api_reference.py`. The stale page documented
  resolved_env_matrix as a keyword argument to classify_compare_pair()
  (that function only accepts request/pair) and omitted the field from
  ResolvedComparePair's own table. `--check` now passes clean.

- tests/test_config_rebalance.py: split the self-contained
  TestEnvironmentMatrixStrictMode class (108 lines) out to a new sibling
  module, tests/test_config_rebalance_env_matrix.py, bringing the original
  file from 1206 to 1096 lines -- back under the architecture gate's
  1200-line test-module cap without a no_growth debt.yaml entry. All 78
  tests (72 + 6) still pass across both files.

Verified: mypy abicheck/, ruff check, ruff format --check (on touched
files), python scripts/check_architecture.py, and
python scripts/gen_python_api_reference.py --check all pass clean.
